common/robovast_common/pyproject.toml
[tool.poetry]
name = "robovast-common"
version = "1.0.0"
description = "Common components for RoboVast"
authors = ["Frederik Pasch <fred-labs@mailbox.org>"]
license = "Apache-2.0"
packages = [{include = "robovast_common"}]

[tool.poetry.dependencies]
python = ">=3.12,<3.14"
pyyaml = "^6.0"
numpy = "^1.26"
pyside6-essentials = "^6.6"
pandas = "^2.2"

[tool.poetry.plugins."robovast.variation_types"]
"ParameterVariationList" = "robovast_common.variation.parameter_variation:ParameterVariationList"
"ParameterVariationDistributionUniform" = "robovast_common.variation.parameter_variation:ParameterVariationDistributionUniform"
"ParameterVariationDistributionGaussian" = "robovast_common.variation.parameter_variation:ParameterVariationDistributionGaussian"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
